fixed xib file for ladderVC
