+++ TennisLadder.xcworkspace/contents.xcworkspacedata
@@ -5,9 +5,6 @@
       location = "group:TennisLadder.xcodeproj">
    </FileRef>
    <FileRef
-      location = "group:TennisLadder">
-   </FileRef>
-   <FileRef
       location = "group:Pods/Pods.xcodeproj">
    </FileRef>
 </Workspace>
